rename top-layer package ota_client to otaclient
