+++ otaclient/app/boot_control/common.py
