+++ otaclient/app/common.py
