+++ otaclient/app/create_standby/common.py
