+++ otaclient/app/create_standby/legacy_mode.py
