+++ otaclient/app/create_standby/rebuild_mode.py
