+++ otaclient/app/ecu_info.py
