+++ otaclient/app/ota_client_stub.py
