+++ otaclient/app/ota_metadata.py
