+++ otaclient/app/update_stats.py
