+++ otaclient/aws_iot_log_server/.gitignore
