+++ otaclient/aws_iot_log_server/app/aws_iot_logger.py
