+++ otaclient/aws_iot_log_server/app/boto3_session.py
