+++ otaclient/aws_iot_log_server/app/custom_http_handler.py
