+++ otaclient/aws_iot_log_server/app/http_server.py
