+++ otaclient/aws_iot_log_server/app/log_proxy_server.py
