+++ otaclient/aws_iot_log_server/tests/data/greengrass/config.json
