+++ otaclient/aws_iot_log_server/tests/test_boto3_session.py
